feat: Add axios dependency and update logout functionality
This commit adds the axios dependency to the project by updating the package.json file. It also modifies the logout functionality in the authService.js file to make a POST request to the '/logout' endpoint using the apiClient. Additionally, it updates the AsyncStorage key used for removing the authentication token. The changes in the HomeScreen.js file include checking the authentication status before fetching products and displaying an alert if the user is not logged in. In the ProfileScreen.js file, the error messages displayed in the Alert component are updated to be in English. Lastly, the AppNavigator.js file is modified to conditionally render screens based on the authentication status.